+++ src/constants/appVersion.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = "13.316.2";
+export const APP_VERSION = "13.317.1";